Mesh and mesh primitives, metalness/roughness factors in materials
Mesh and Materials cleanup, also added mesh primitives to mirror the way GLTF stores meshes and its primitives. Moved the default white texture to the renderer and also using it for the metallic roughness texture if missing. Added a default normal texture to the renderer if missing.
Removed the color attribute from vertex instance data, and added the metalness/roughness factors in the instance data.