RenderResourceHandle implementation
Finished the implementation for RenderResourceHandles, the renderer stores buffers, textures, meshes, and materials in slotmaps now and only returns a RenderResourceHandle to the outside.
A new bug has been introduced which crashes on resize because of the reallocation of the render target descriptors.
Mesh component ImGui rendering is currently disabled.