PCF 5x5 grid, crude gaussian blur (The Witness)
Implemented PCF sampling with a 5x5 grid and a crude gaussian blur, using 9 PCF samples in total and weighing them accordingly. The source for this implementation was the shadow mapping implementation in "The Witness".
Moved code for calculating shadows into its own file.
Moved code for evaluating the BRDF into its own file.