Halton Jitter, TAA history, TAA preparations
Implemented per-frame jitter using a halton sequence.
Added a history buffer which will contain the TAA resolve history.
Changed the tonemap settings to be part of the new global constant buffer, which has a bunch of settings for rendering the current frame.
Added a Temporal anti-aliasing compute shader.
Added GPU pointer to the DescriptorAllocation class.